added again validations for manager and nation

+++ Controllers/ManagerController.cs
@@ -15,6 +15,10 @@ namespace FutManager.Controllers
         }
         public RedirectToActionResult Add(string FirstName, string LastName,  int NationalityId, int ClubId, int Age, int Rating, bool isReal)
         {
+            if (FirstName == null || LastName == null || Rating < 1  || Age < 1 || !DataService.GetClubs().Any(y => y.Id == ClubId) || ClubId == 0 || NationalityId == 0 || !DataService.GetNations().Any(y => y.Id == NationalityId))
+            {
+                return RedirectToAction(actionName: "Index", controllerName: "Error");
+            }
             DataService.AddManager(FirstName, LastName, Age , NationalityId, ClubId, Rating, isReal);
             return RedirectToAction(actionName: "Index");
         }
@@ -33,6 +37,10 @@ namespace FutManager.Controllers
         }
         public IActionResult Details(int id)
         {
+            if (!DataService.GetClubs().Any(y => y.Id == id)  || !DataService.GetNations().Any(y => y.Id == id))
+            {
+                return RedirectToAction(actionName: "Index", controllerName: "Error");
+            }
             List<Club> clubs = DataService.GetClubs();
             List<Nation> nations = DataService.GetNations();
             Manager manager = DataService.GetManagers().FirstOrDefault(x => x.Id == id);
@@ -44,6 +52,11 @@ namespace FutManager.Controllers
 
         public IActionResult Delete(int id)
         {
+
+            if (!DataService.GetClubs().Any(y => y.Id == id) || id == 0 || !DataService.GetNations().Any(y => y.Id == id))
+            {
+                return RedirectToAction(actionName: "Index", controllerName: "Error");
+            }
             Manager manager = DataService.GetManagers().FirstOrDefault(y => y.Id == id);
             return View(manager);
         }
@@ -52,6 +65,10 @@ namespace FutManager.Controllers
         {
             if (password == "password")
             {
+                if (!DataService.GetClubs().Any(y => y.Id == id) || id == 0 || !DataService.GetNations().Any(y => y.Id == id))
+                {
+                    return RedirectToAction(actionName: "Index", controllerName: "Error");
+                }
                 DataService.DeleteManager(id);
             }
             return RedirectToAction(actionName: "Index");
@@ -60,7 +77,10 @@ namespace FutManager.Controllers
         public IActionResult Edit(int id)
         {
             ManagerAndClubsAndNations x = new ManagerAndClubsAndNations();
-
+            if (!DataService.GetClubs().Any(y => y.Id == id) || id == 0 || !DataService.GetNations().Any(y => y.Id == id))
+            {
+                return RedirectToAction(actionName: "Index", controllerName: "Error");
+            }
             x.manager = DataService.GetManagers().FirstOrDefault(y => y.Id == id);
             x.Clubs = DataService.GetClubs();
             x.Nations = DataService.GetNations();
@@ -70,7 +90,12 @@ namespace FutManager.Controllers
 
         public RedirectToActionResult ConfirmEdit(int id, string FirstName, string LastName, int NationalityId, int ClubId, int Age, int Rating, bool isReal,string password)
         {
-            if (password == "password") {
+            if (password == "password") 
+            {
+                if (FirstName == null || LastName == null || Rating < 1 || Age < 1 || !DataService.GetClubs().Any(y => y.Id == ClubId) || ClubId == 0 || NationalityId == 0 || !DataService.GetNations().Any(y => y.Id == NationalityId))
+                {
+                    return RedirectToAction(actionName: "Index", controllerName: "Error");
+                }
                 DataService.EditManager(id, FirstName, LastName, NationalityId, ClubId, Age, Rating, isReal);
             }
             return RedirectToAction(actionName: "Index");
